changed internal names and setting entry's
Changed to a more easier way of how they get filled and refilled.
Cleaned up the local file and the setting file. Renamed a setting name to make it more clear for what it is and to make it easier for the future editing. 
Added the missing English locale entries and translated them.
Tested Both English and German to ensure it is working correctly.

This is the last commit I am doing for this year, I wish you happy Holidays and merry Christmas.